+++ src/ska_oso_services/pht/__init__.py
@@ -9,7 +9,7 @@ from ska_oso_services.pht.api import (
     reviews,
 )
 
-router = APIRouter(prefix="/pht", tags=["PHT API"])
+router = APIRouter(prefix="/pht")
 router.include_router(prsls.router)
 router.include_router(reviewers.router)
 router.include_router(reviews.router)