+++ tests/component/test_panel_decision_api.py
@@ -10,13 +10,11 @@ import json
 import uuid
 from http import HTTPStatus
 
-import requests
-
 from ..unit.util import VALID_PANEL_DECISION, TestDataFactory
 from . import PHT_URL
 
 
-def test_create_and_get_panel_decision():
+def test_create_and_get_panel_decision(authrequests):
     """
     Integration test for the POST /panel-decision/ endpoint
     and GET /panel-decision/{decision_id}.
@@ -24,7 +22,7 @@ def test_create_and_get_panel_decision():
     """
 
     # POST using JSON string
-    post_response = requests.post(
+    post_response = authrequests.post(
         f"{PHT_URL}/panel-decisions/",
         data=VALID_PANEL_DECISION,
         headers={"Content-Type": "application/json"},
@@ -36,7 +34,7 @@ def test_create_and_get_panel_decision():
     ), f"Expected string, got {type(decision_id)}: {decision_id}"
 
     # GET created proposal
-    get_response = requests.get(f"{PHT_URL}/panel-decisions/{decision_id}")
+    get_response = authrequests.get(f"{PHT_URL}/panel-decisions/{decision_id}")
     assert get_response.status_code == HTTPStatus.OK, get_response.content
     actual_payload = get_response.json()
 
@@ -52,7 +50,7 @@ def test_create_and_get_panel_decision():
     assert actual_payload == expected_payload
 
 
-def test_panel_decision_create_then_put():
+def test_panel_decision_create_then_put(authrequests):
     """
     Test that an entity POSTed to /panel-decision/create
     can then be updated with PUT /panel-decision/{identifier},
@@ -60,7 +58,7 @@ def test_panel_decision_create_then_put():
     """
 
     # POST a new proposal
-    post_response = requests.post(
+    post_response = authrequests.post(
         f"{PHT_URL}/panel-decisions/",
         data=VALID_PANEL_DECISION,
         headers={"Content-Type": "application/json"},
@@ -74,7 +72,7 @@ def test_panel_decision_create_then_put():
     assert returned_decision_id == expected_decision_id
 
     # GET proposal to fetch latest state
-    get_response = requests.get(f"{PHT_URL}/panel-decisions/{returned_decision_id}")
+    get_response = authrequests.get(f"{PHT_URL}/panel-decisions/{returned_decision_id}")
     assert get_response.status_code == HTTPStatus.OK, get_response.content
 
     originalreview = get_response.json()
@@ -84,7 +82,7 @@ def test_panel_decision_create_then_put():
     review_to_update = json.dumps(originalreview)
 
     # PUT updated proposal
-    put_response = requests.put(
+    put_response = authrequests.put(
         f"{PHT_URL}/panel-decisions/{returned_decision_id}",
         data=review_to_update,
         headers={"Content-Type": "application/json"},
@@ -94,7 +92,7 @@ def test_panel_decision_create_then_put():
     assert put_response.json()["metadata"]["version"] == initial_version + 1
 
 
-def test_get_list_panel_decision_for_user():
+def test_get_list_panel_decision_for_user(authrequests):
     """
     Integration test:
     - Create multiple reviews
@@ -111,7 +109,7 @@ def test_get_list_panel_decision_for_user():
         proposal = TestDataFactory.panel_decision(decision_id=decision_id)
         proposal_json = proposal.model_dump_json()
 
-        response = requests.post(
+        response = authrequests.post(
             f"{PHT_URL}/panel-decisions/",
             data=proposal_json,
             headers={"Content-Type": "application/json"},
@@ -121,12 +119,12 @@ def test_get_list_panel_decision_for_user():
 
     # Get created_by from one of the created reviews
     example_decision_id = created_ids[0]
-    get_response = requests.get(f"{PHT_URL}/panel-decisions/{example_decision_id}")
+    get_response = authrequests.get(f"{PHT_URL}/panel-decisions/{example_decision_id}")
     assert get_response.status_code == HTTPStatus.OK, get_response.content
     user_id = get_response.json()["metadata"]["created_by"]
 
     # GET /list/{user_id}
-    list_response = requests.get(f"{PHT_URL}/panel-decisions/list/{user_id}")
+    list_response = authrequests.get(f"{PHT_URL}/panel-decisions/list/{user_id}")
     assert list_response.status_code == HTTPStatus.OK, list_response.content
 
     reviews = list_response.json()