+++ tests/component/test_panels.py
@@ -1,8 +1,5 @@
-import json
 import uuid
 from http import HTTPStatus
-
-import requests
 
 from ..unit.util import REVIEWERS, TestDataFactory
 from . import PHT_URL
@@ -11,11 +8,11 @@ PANELS_API_URL = f"{PHT_URL}/panels"
 HEADERS = {"Content-type": "application/json"}
 
 
-def test_create_panel():
+def test_create_panel(authrequests):
     panel = TestDataFactory.panel_basic()
     data = panel.json()
 
-    response = requests.post(f"{PANELS_API_URL}", data=data, headers=HEADERS)
+    response = authrequests.post(f"{PANELS_API_URL}", data=data, headers=HEADERS)
 
     assert response.status_code == HTTPStatus.OK
 
@@ -23,13 +20,13 @@ def test_create_panel():
     assert panel.panel_id == result
 
 
-def test_panels_post_duplicate_reviewer():
+def test_panels_post_duplicate_reviewer(authrequests):
     panel = TestDataFactory.panel()
     panel.reviewers.append(panel.reviewers[0])
 
     data = panel.json()
 
-    response = requests.post(f"{PANELS_API_URL}", data=data, headers=HEADERS)
+    response = authrequests.post(f"{PANELS_API_URL}", data=data, headers=HEADERS)
 
     assert response.status_code == HTTPStatus.CONFLICT
 
@@ -38,13 +35,13 @@ def test_panels_post_duplicate_reviewer():
     assert expected == result
 
 
-def test_panels_post_duplicate_proposal():
+def test_panels_post_duplicate_proposal(authrequests):
     panel = TestDataFactory.panel(reviewer_id=REVIEWERS[0]["id"])
     panel.proposals.append(panel.proposals[0])
 
     data = panel.json()
 
-    response = requests.post(f"{PANELS_API_URL}", data=data, headers=HEADERS)
+    response = authrequests.post(f"{PANELS_API_URL}", data=data, headers=HEADERS)
 
     assert response.status_code == HTTPStatus.CONFLICT
 
@@ -55,12 +52,12 @@ def test_panels_post_duplicate_proposal():
     assert expected == result
 
 
-def test_panels_post_not_existing_reviewer():
+def test_panels_post_not_existing_reviewer(authrequests):
     panel = TestDataFactory.panel()
 
     data = panel.json()
 
-    response = requests.post(f"{PANELS_API_URL}", data=data, headers=HEADERS)
+    response = authrequests.post(f"{PANELS_API_URL}", data=data, headers=HEADERS)
 
     assert response.status_code == HTTPStatus.BAD_REQUEST
 
@@ -69,12 +66,12 @@ def test_panels_post_not_existing_reviewer():
     assert expected == result
 
 
-def test_panels_post_not_existing_proposal():
+def test_panels_post_not_existing_proposal(authrequests):
     panel = TestDataFactory.panel(reviewer_id=REVIEWERS[0]["id"])
 
     data = panel.json()
 
-    response = requests.post(f"{PANELS_API_URL}", data=data, headers=HEADERS)
+    response = authrequests.post(f"{PANELS_API_URL}", data=data, headers=HEADERS)
 
     assert response.status_code == HTTPStatus.BAD_REQUEST
 
@@ -83,49 +80,7 @@ def test_panels_post_not_existing_proposal():
     assert expected == result
 
 
-def test_get_reviews_for_panel_with_wrong_id():
-    panel_id = "wrong id"
-    response = requests.get(f"{PHT_URL}/panels/reviews/{panel_id}")
-    assert response.status_code == HTTPStatus.OK
-    res = response.json()
-    assert [] == res
-
-
-def test_get_reviews_for_panel_with_valid_id():
-    proposal = TestDataFactory.complete_proposal("my proposal")
-    response = requests.post(
-        f"{PHT_URL}/prsls/create", data=proposal.json(), headers=HEADERS
-    )
-    assert response.status_code == HTTPStatus.OK
-
-    panel_id = "panel-test-20250717-00001"
-    panel = TestDataFactory.panel_basic(panel_id=panel_id, name="New name")
-    data = panel.json()
-    response = requests.post(f"{PANELS_API_URL}", data=data, headers=HEADERS)
-    assert response.status_code == HTTPStatus.OK
-
-    review = TestDataFactory.reviews(
-        panel_id=panel_id,
-        review_id="my review id",
-        prsl_id="my proposal",
-        reviewer_id=REVIEWERS[0]["id"],
-    )
-    response = requests.post(f"{PHT_URL}/reviews/", data=review.json(), headers=HEADERS)
-    assert response.status_code == HTTPStatus.OK
-
-    response = requests.get(f"{PHT_URL}/panels/reviews/{panel_id}")
-    assert response.status_code == HTTPStatus.OK
-    res = response.json()
-    del res[0]["metadata"]
-
-    review = json.loads(review.json())
-    del review["metadata"]
-    expected = [review]
-
-    assert expected == res
-
-
-def test_get_list_panels_for_user():
+def test_get_list_panels_for_user(authrequests):
     """
     Integration test:
     - Create multiple panels
@@ -142,7 +97,7 @@ def test_get_list_panels_for_user():
         panel = TestDataFactory.panel_basic(panel_id=panel_id, name=f"Star{i+1}")
         panel_json = panel.model_dump_json()
 
-        response = requests.post(
+        response = authrequests.post(
             f"{PANELS_API_URL}",
             data=panel_json,
             headers={"Content-Type": "application/json"},
@@ -152,12 +107,12 @@ def test_get_list_panels_for_user():
 
     # Get created_by from one of the created panels
     example_panel_id = created_ids[0]
-    get_response = requests.get(f"{PANELS_API_URL}/{example_panel_id}")
+    get_response = authrequests.get(f"{PANELS_API_URL}/{example_panel_id}")
     assert get_response.status_code == HTTPStatus.OK, get_response.content
     user_id = get_response.json()["metadata"]["created_by"]
 
     # GET /list/{user_id}
-    list_response = requests.get(f"{PANELS_API_URL}/list/{user_id}")
+    list_response = authrequests.get(f"{PANELS_API_URL}/list/{user_id}")
     assert list_response.status_code == HTTPStatus.OK, list_response.content
 
     panels = list_response.json()